Remove the rating period. Compute new rating instantly.

+++ backend/ranking/rating.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass
 class Rating:
-    value: int
-    deviation: int
+    value: float
+    deviation: float
